refactor : 패키지 명 member에서 user로 변경
